Add star ratings to film
