+++ static/src/javascripts/projects/common/modules/experiments/ab.js
@@ -10,7 +10,7 @@ define([
     'common/modules/experiments/tests/live-blog-chrome-notifications-prod',
     'common/modules/experiments/tests/loyal-adblocking-survey',
     'common/modules/experiments/tests/facebook-share-params',
-    'common/modules/experiments/tests/participation-star-ratings',
+    'common/modules/experiments/tests/participation-low-fric-film',
     'common/modules/experiments/tests/participation-low-fric-music-v1',
     'common/modules/experiments/tests/participation-low-fric-tv',
     'common/modules/experiments/tests/participation-low-fric-recipes',
@@ -45,7 +45,7 @@ define([
     LiveBlogChromeNotificationsProd,
     LoyalAdblockingSurvey,
     FacebookShareParams,
-    ParticipationStarRatings,
+    ParticipationLowFricFilm,
     ParticipationLowFricMusicV1,
     ParticipationLowFricTv,
     ParticipationLowFricRecipes,
@@ -76,7 +76,7 @@ define([
         new LiveBlogChromeNotificationsProd(),
         new LoyalAdblockingSurvey(),
         new FacebookShareParams(),
-        new ParticipationStarRatings(),
+        new ParticipationLowFricFilm(),
         new ParticipationLowFricMusicV1(),
         new ParticipationLowFricTv(),
         new ParticipationLowFricRecipes(),
